+++ src/main/java/com/handwoong/rainbowletter/common/config/schedule/ScheduleConfig.java
@@ -1,0 +1,9 @@
+package com.handwoong.rainbowletter.common.config.schedule;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+@Configuration
+@EnableScheduling
+public class ScheduleConfig {
+}
