--- src/main/java/com/handwoong/rainbowletter/gpt/controller/port/ChatGptService.java
@@ -1,4 +1,0 @@
-package com.handwoong.rainbowletter.gpt.controller.port;
-
-public interface ChatGptService {
-}
