--- src/main/java/com/handwoong/rainbowletter/gpt/service/port/ChatGptRepository.java
@@ -1,4 +1,0 @@
-package com.handwoong.rainbowletter.gpt.service.port;
-
-public interface ChatGptRepository {
-}
