+++ src/main/java/com/handwoong/rainbowletter/letter/controller/port/ReplyService.java
@@ -1,4 +1,5 @@
 package com.handwoong.rainbowletter.letter.controller.port;
 
 public interface ReplyService {
+    void create();
 }
