+++ src/main/java/com/handwoong/rainbowletter/letter/infrastructure/ChatGptJpaRepository.java
@@ -1,4 +1,4 @@
-package com.handwoong.rainbowletter.gpt.infrastructure;
+package com.handwoong.rainbowletter.letter.infrastructure;
 
 import org.springframework.data.jpa.repository.JpaRepository;
