+++ src/main/java/com/handwoong/rainbowletter/letter/infrastructure/ChatGptRepositoryImpl.java
@@ -1,6 +1,6 @@
-package com.handwoong.rainbowletter.gpt.infrastructure;
+package com.handwoong.rainbowletter.letter.infrastructure;
 
-import com.handwoong.rainbowletter.gpt.service.port.ChatGptRepository;
+import com.handwoong.rainbowletter.letter.service.port.ChatGptRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
