+++ src/main/java/com/handwoong/rainbowletter/letter/service/ReplyServiceImpl.java
@@ -1,8 +1,10 @@
 package com.handwoong.rainbowletter.letter.service;
 
 import com.handwoong.rainbowletter.letter.controller.port.ReplyService;
+import com.handwoong.rainbowletter.letter.service.port.LetterRepository;
 import com.handwoong.rainbowletter.letter.service.port.ReplyRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -10,5 +12,13 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class ReplyServiceImpl implements ReplyService {
+    private final LetterRepository letterRepository;
     private final ReplyRepository replyRepository;
+
+    @Override
+    @Scheduled(cron = "0 0 * * * *", zone = "Asia/Seoul")
+    @Transactional
+    public void create() {
+
+    }
 }
