+++ src/main/java/com/handwoong/rainbowletter/letter/service/port/ChatGptRepository.java
@@ -1,0 +1,4 @@
+package com.handwoong.rainbowletter.letter.service.port;
+
+public interface ChatGptRepository {
+}
